+++ src/main/java/com/uisrael/drinkhouse/presentacion/dto/response/ProveedorResponseDto.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class ProveedorResponseDto {
 
-	private Integer proveedorId;
+	private Long proveedorId;
 	private String ruc;
 	private String razonSocial;
 	private String direccion;
